Test and lint (#3)
* Added chaining to base test cases.

* Added basic integration tests to newman.

* Fixed unique constraints on DB

* Added regression testing using newman.

* Unit testing for serializer utilities.

* Added local linting

* Added static code analyzer.

* Integrated Winston middleware for logging.

* Post lint change

* Added documentation.

+++ server/controllers/playersController.js
@@ -1,5 +1,8 @@
-import { Player, ManagementRelationship } from "../models/index";
-import playerSerializer from "../serializers/playerSerializer";
+import Sequelize from 'sequelize';
+
+import { Player, ManagementRelationship } from '../models/index';
+import playerSerializer from '../serializers/playerSerializer';
+import { logger } from '../config/loggerConfig';
 
 const index = async (req, res) => {
   try {
@@ -23,6 +26,7 @@ const index = async (req, res) => {
       players: playersArray
     });
   } catch (err) {
+    logger.error(err);
     return res.status(400).json({ success: false });
   }
 };
@@ -48,31 +52,48 @@ const create_player = async (req, res) => {
       player: playerCreated
     });
   } catch (err) {
+    logger.error(err);
     return res.status(400).json({ success: false });
   }
 };
 
 const delete_player = async (req, res) => {
   try {
-    const player_records_destroyed = await Player.destroy({
+    const playerUUID = req.params.id;
+    const adminUUID = req.body.user.id;
+    const Op = Sequelize.Op;
+
+    const managementRecord = await ManagementRelationship.findOne({
       where: {
-        id: req.params.id
+        adminUUID,
+        [Op.and]: {
+          playerUUID
+        }
       }
     });
 
-    if (player_records_destroyed == 1) {
-      const relationship_records_destroyed = await ManagementRelationship.destroy(
-        {
-          where: {
-            playerUUID: req.params.id
-          }
+    if (managementRecord) {
+      const player_records_destroyed = await Player.destroy({
+        where: {
+          id: req.params.id
         }
-      );
+      });
+
+      if (player_records_destroyed == 1) {
+        const relationship_records_destroyed = await ManagementRelationship.destroy(
+          {
+            where: {
+              playerUUID: req.params.id
+            }
+          }
+        );
+      }
     } else {
-      throw "Error in record destruction";
+      throw new Error('Error in record destruction');
     }
     return res.json({ success: true });
   } catch (err) {
+    logger.error(err);
     return res.status(400).json({ success: false });
   }
 };
